Allow for jest globals being missing
It is valid for a jest configuration to disable globals being injected, which prevents from importing this package. The proposed change here keeps the option to enforce globals are present, but changes the default behavior to allowing the globals to be absent. Enforcing their presence relies on an additional function call. Preserving the current default behaviour could be done by setting an environment variable if the default behavior change isn't acceptable.